[amnn/lily58] Swap main board
Making the right hand side (containing the trackball) makes it trivial
to set-up the trackball (if left were main, the trackball state would
need to be communicated across the serial connection).

Also made the OLED rendering code agnostic of main vs not, and pin the
detail to the left side.