[lily58/amnn] Static imagine on non-main OLED
Now I understand why the OLED code checks whether it's running on the
main keyboard or not:  It has the same problem as the mouse that only
the main keyboard tracks the state that it was displaying (keypresses
for the keylogger as well as all the modifier key states).

Set the OLED to a static image for now, while we figure out what to do
with it.